Added weapon recoil
Finally got it to work well enough. Would have liked to have made the camera return to the center when done firing instead of to the mouse position, since it can be quite annoying having to move the mouse back up after firing for a longer period. We'll see if I will fix it in the future